OTR(Frontend): Renamings
