+++ frontend/packages/otr/src/components/publicInterpreter/listing/row/CollapsibleRow.tsx
